backend/requirements.txt
Flask==2.3.2
paho-mqtt==1.6.1
pymongo==4.5.0
flask-cors==4.0.0
